+++ vertico-prescient.el
@@ -8,7 +8,7 @@
 ;; Created: 23 Sep 2022
 ;; Package-Requires: ((emacs "27.1") (prescient "6.1.0") (vertico "0.28"))
 ;; SPDX-License-Identifier: MIT
-;; Version: 6.1.0
+;; Version: 6.2.0
 
 ;;; Commentary: